refactor: extract web client state modules
