+++ openspec/changes/archive/2026-08-21-decompose-web-send-message/.openspec.yaml
@@ -1,0 +1,3 @@
+schema: spec-driven
+created: 2026-08-21
+skip_specs: true
