+++ openspec/changes/archive/2026-08-21-extract-web-client-state-modules/proposal.md
