+++ openspec/changes/archive/archive-index.json
@@ -3085,6 +3085,28 @@
         "multi-agent-group-chat",
         "session-runtime-management"
       ]
+    },
+    {
+      "archivedOn": "2026-08-21",
+      "changeName": "decompose-web-send-message",
+      "path": "openspec/changes/archive/2026-08-21-decompose-web-send-message/",
+      "artifacts": [
+        "proposal",
+        "design",
+        "tasks"
+      ],
+      "capabilities": []
+    },
+    {
+      "archivedOn": "2026-08-21",
+      "changeName": "extract-web-client-state-modules",
+      "path": "openspec/changes/archive/2026-08-21-extract-web-client-state-modules/",
+      "artifacts": [
+        "proposal",
+        "design",
+        "tasks"
+      ],
+      "capabilities": []
     }
   ]
 }
